back/recognition/pyproject.toml
[tool.poetry]
name = "kanji-recognition"
version = "0.1.0"
description = ""
authors = ["Alexandre <em_alexandre@yahoo.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.3"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
gunicorn = "^21.2.0"
tensorflow = "^2.13.0"
python-multipart = "^0.0.6"
pyjwt = "^2.8.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
